+++ WebContent/user/login/info.jsp
@@ -1,5 +1,4 @@
-<%@page import="com.deco.login.userDTO"%>
-<%@page import="com.deco.login.loginDAO"%>
+<%@page import="com.deco.user.userDTO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
@@ -59,7 +58,7 @@
 				<td><%=udto.getPoint() %></td>
 			</tr>
 		</table>
-	<a href="./main.use">메인 페이지</a> | <a href="./update.use?user_num=<%=udto.getUser_num() %>">정보수정</a>
+	<a href="./main.us">메인 페이지</a> | <a href="./update.us?user_num=<%=udto.getUser_num() %>">정보수정</a>
 	</center>
