+++ WebContent/user/login/update.jsp
@@ -1,4 +1,4 @@
-<%@page import="com.deco.login.userDTO"%>
+<%@page import="com.deco.user.userDTO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
