+++ src/com/deco/user/listAction.java
@@ -1,4 +1,4 @@
-package com.deco.login;
+package com.deco.user;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -6,32 +6,25 @@ import javax.servlet.http.HttpSession;
 
 import com.deco.Action;
 import com.deco.ActionForward;
-import com.deco.user.userDTO;
 
-
-public class UserInfoAction implements Action{
+public class listAction implements Action{
 
 	@Override
 	public ActionForward execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-		
 		HttpSession session = req.getSession();
 		Integer user_num = Integer.parseInt(req.getParameter("user_num"));
-		
-		
 		ActionForward forward = new ActionForward();
-		if(user_num == null){
-			forward = new ActionForward("./login.use", true);
-			return forward;
-		}
-		loginDAO udao = new loginDAO();
-		userDTO udto = udao.info(user_num);
+		
+		userDAO loDAO = new userDAO();
+		userDTO udto = loDAO.info(user_num);
 		
 		req.setAttribute("udto", udto);
 		
-		forward = new ActionForward("./user/login/info.jsp", false);
+		forward = new ActionForward("./user/login/update.jsp", false);
+		
+		// 현재 updateAction 페이지 리스트 페이지로 수정후 updateAction페이지 생성하여 정보 수정하게하는 객체 생성하여 정보수정
+		
 		
 		return forward;
 	}
-
-	
 }
