+++ src/com/deco/user/userController.java
@@ -30,6 +30,33 @@ public class userController extends Controller{
 		}else if(command.equals("/changeEmailCode.us")){
 			action = new changeEmailAction();
 			forward = action.execute(req, res);
+		
+		}else if(command.equals("/login.us")){
+			forward = new ActionForward("./user/login/login.jsp", false);
+		
+		}else if(command.equals("/LoginAction.us")){
+				action = new LoginAction();
+				forward = action.execute(req, res);
+		
+		}else if(command.equals("/main.us")){
+				forward = new ActionForward("./main/main.jsp",false);
+		
+		}else if(command.equals("/info.us")){
+				System.out.println("info.us 호출");
+				action = new UserInfoAction();
+				forward = action.execute(req, res);
+		
+		}else if(command.equals("/userlogout.us")){
+			action = new userLogoutAction();
+				forward = action.execute(req, res);
+		
+		}else if(command.equals("/update.us")){
+				action = new listAction();
+				forward = action.execute(req, res);
+		
+		}else if(command.equals("/UpdateAction.us")){
+				action = new UpdateAction();
+				forward = action.execute(req, res);
 		}
 		
 		render(forward,req,res);
